Fix stream not being flushed
Currently this transport only sends the first log to Rollbar. All subsequent logs entirely bypass this transport. This happens because `callback` is never called because the `cb` is not bound to the current `this`.

This regression was introduced in cd6225966 when `cb` was converted from an arrow function (implicit binding) to a function expression (unbound).